+++ packages/wethegit-components-cli/src/commands/add/index.ts
@@ -1,13 +1,19 @@
-import { resolve } from "node:path";
-import ora from "ora";
 import chalk from "chalk";
-import fse from "fs-extra";
-import appRootPath from "app-root-path";
-import resolvePackagePath from "resolve-package-path";
 
-import { logger, buildAndParseConfig, ensureCwd } from "../../utils";
+import {
+  logger,
+  buildAndParseConfig,
+  ensureCwd,
+  handleError,
+  ensureComponentsPackageIsInstalled,
+} from "../../utils";
 
-import { copyComponentsByName, promptForComponents } from "./utils";
+import {
+  copyComponentsByName,
+  formatComponentFilesWithPrettier,
+  promptForComponents,
+} from "./utils";
+import { resolve } from "path";
 
 interface Options {
   root: string;
@@ -15,73 +21,56 @@ interface Options {
 
 export async function add(options: Options) {
   try {
-    let componentsPackageRoot = resolvePackagePath(
-      "@wethegit/components",
-      appRootPath.toString()
-    );
+    ensureComponentsPackageIsInstalled();
 
-    if (!componentsPackageRoot) {
-      logger.error(
-        "Failed to find @wethegit/components package. It should be installed as a dependency automatically, something went wrong with the CLI installation."
-      );
-      process.exit(1);
-    }
-
-    componentsPackageRoot = resolve(componentsPackageRoot, "../src");
     const { root } = options;
 
     // different than init if cwd doesnt exist this will throw, user should run init first to ensure deps and other requirements are met
     const cwd = await ensureCwd(root);
 
     // get our config
-    const config = await buildAndParseConfig(cwd);
+    const config = await buildAndParseConfig(cwd, { errorIfNotFound: true });
 
     // ask what components to install
     const { selectedComponentNames, proceed } = await promptForComponents();
 
     if (!proceed || selectedComponentNames.length <= 0) process.exit(1);
 
-    // create components dir if it doesnt exist
-    if (!(await fse.pathExists(config.componentsRootDir))) {
-      const { componentsRootDir } = config;
-
-      const spinner = ora(
-        `Components root directory ${chalk.cyan(
-          componentsRootDir
-        )} does not exist, creating...`
-      )?.start();
-
-      try {
-        await fse.ensureDir(componentsRootDir);
-        await spinner.succeed();
-      } catch (e) {
-        await spinner.fail(
-          `Error creating ${chalk.cyan(componentsRootDir)} directory`
-        );
-        logger.error(e);
-        process.exit(1);
-      }
-    }
-
     // copy components
     try {
       await copyComponentsByName({
-        componentsPackageRoot,
+        cwd,
         config,
         selectedComponentNames,
       });
-    } catch (e) {
-      logger.error("");
-      logger.error(e);
-      process.exit(1);
+    } catch (error) {
+      handleError({
+        error,
+        exit: true,
+      });
+    }
+
+    // format component files
+    try {
+      await formatComponentFilesWithPrettier({
+        config,
+        selectedComponentNames,
+        cwd,
+      });
+    } catch (error) {
+      handleError({
+        error,
+        exit: true,
+      });
     }
 
     logger.info("");
     logger.success(`${chalk.green("Success!")} All done!`);
     logger.info("");
   } catch (error) {
-    logger.error("");
-    logger.error(error);
-    process.exit(0);
+    handleError({
+      error,
+      exit: true,
+    });
   }
 }