+++ packages/wethegit-components-cli/src/commands/add/utils/transformTsToJs.ts
@@ -5,18 +5,34 @@ import fse from "fs-extra";
 import { logger } from "../../../utils";
 
 interface TransformTsToJsOptions {
-  dest: string;
+  cwd: string;
+  destDir: string;
   files: string[];
 }
 
-export async function transformTsToJs({ dest, files }: TransformTsToJsOptions) {
+/**
+ * Transforms ts files to js files. If a file is not a ts file, it will be copied directly.
+ */
+export async function transformTsToJs({
+  cwd,
+  destDir,
+  files,
+}: TransformTsToJsOptions) {
+  const defaultTsConfigPath = resolve(cwd, "./tsconfig.json");
+  const isThereATsConfig = await fse.pathExists(defaultTsConfigPath);
+
   const tsProject = new Project({
+    ...(isThereATsConfig
+      ? { tsConfigFilePath: defaultTsConfigPath }
+      : { undefined }),
     skipAddingFilesFromTsConfig: true,
     skipFileDependencyResolution: true,
     compilerOptions: {
       target: ScriptTarget.ESNext,
-      outDir: dest,
+      outDir: destDir,
       jsx: ts.JsxEmit.Preserve,
+      declaration: true,
+      declarationDir: resolve(cwd, "types"),
     },
   });
 
@@ -25,7 +41,7 @@ export async function transformTsToJs({ dest, files }: TransformTsToJsOptions) {
 
   for (let file of files) {
     const filename = basename(file);
-    const outFile = resolve(dest, filename);
+    const outFile = resolve(destDir, filename);
 
     if ([".ts", ".tsx"].includes(extname(file))) {
       try {